Add 3D stuff
MSVC errors may still occur

+++ src/renderer/vert_array.h
@@ -2,7 +2,6 @@
 #define R_VERTEX_ARRAY_H
 
 #include "util/types.h"
-#include <stddef.h>
 
 typedef struct 
 { 
@@ -11,7 +10,7 @@ typedef struct
 
 void va_create(vert_array *va);
 void va_bind(vert_array *va);
-void va_add_attrib(uint index, uint count, uint type, boolean normalize, size_t stride, const void *offset);
+void va_add_attrib(uint index, uint count, uint type, boolean normalize, int stride, const void *offset);
 void va_unbind();
 void va_destroy(vert_array *va);
